pyproject.toml
[project]
name = "intel-image-cl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "seaborn>=0.13.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.1.2",
    "flake8-pyproject>=1.2.3",
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "jupyterlab>=4.3.6",
    "notebook>=7.3.3",
]

[tool.flake8]
use_pyproject = true
max-line-length = 90
ignore = ['E203', 'E501', 'W503', 'B950']
max-complexity = 12
select = ['B', 'C', 'E', 'F', 'W', 'B9']
per-file-ignores = [
    # for easier imports to __init__ without __all__
    '**/__init__.py: F401',
    # file to define custom types
    'types.py: F401',
]
count = true
